Move more classes and JSPs to specimen module
