+++ api/src/org/labkey/api/study/assay/SpecimenForeignKey.java
@@ -280,9 +280,9 @@ public class SpecimenForeignKey extends LookupForeignKey
 
         DetailsURL detailsURL;
         if (targetStudy != null)
-            detailsURL = DetailsURL.fromString("specimen-specimenEventsRedirect.view?id=${" + parent.getFieldKey() + "}&targetStudy=" + targetStudy.getId() );
+            detailsURL = DetailsURL.fromString("specimen2-specimenEventsRedirect.view?id=${" + parent.getFieldKey() + "}&targetStudy=" + targetStudy.getId() );
         else
-            detailsURL = DetailsURL.fromString("specimen-specimenEventsRedirect.view?id=${" + parent.getFieldKey() + "}&targetStudy=${" + targetStudyFK + "}");
+            detailsURL = DetailsURL.fromString("specimen2-specimenEventsRedirect.view?id=${" + parent.getFieldKey() + "}&targetStudy=${" + targetStudyFK + "}");
         return detailsURL;
     }
