+++ specimen/src/org/labkey/specimen/pipeline/SpecimenPipeline.java
@@ -1,16 +1,16 @@
-package org.labkey.study.pipeline;
+package org.labkey.specimen.pipeline;
 
 import org.labkey.api.module.Module;
 import org.labkey.api.pipeline.PipeRoot;
 import org.labkey.api.pipeline.PipelineDirectory;
 import org.labkey.api.pipeline.PipelineProvider;
 import org.labkey.api.security.permissions.InsertPermission;
-import org.labkey.api.specimen.pipeline.SpecimenBatch;
+import org.labkey.api.specimen.pipeline.AbstractSpecimenTask;
 import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.SpecimenTransform;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.view.ViewContext;
-import org.labkey.study.controllers.specimen.SpecimenController;
+import org.labkey.specimen.action.SpecimenController2.ImportSpecimenDataAction;
 
 import java.io.File;
 
@@ -21,11 +21,10 @@ public class SpecimenPipeline extends PipelineProvider
         super("Specimen", owningModule);
     }
 
-
     @Override
     public void updateFileProperties(final ViewContext context, PipeRoot pr, PipelineDirectory directory, boolean includeAll)
     {
-        if (SpecimenService.get() == null ||
+        if (
             !context.getContainer().hasPermission(context.getUser(), InsertPermission.class) ||
             context.getContainer().isDataspace() || // Cannot import specimens into Dataspace container
             StudyService.get().getStudy(context.getContainer()) == null)
@@ -38,7 +37,7 @@ public class SpecimenPipeline extends PipelineProvider
             @Override
             public boolean accept(File f)
             {
-                if (SpecimenBatch.ARCHIVE_FILE_TYPE.isType(f))
+                if (AbstractSpecimenTask.ARCHIVE_FILE_TYPE.isType(f))
                     return true;
                 else
                 {
@@ -52,8 +51,7 @@ public class SpecimenPipeline extends PipelineProvider
             }
         });
 
-        String actionId = createActionId(SpecimenController.ImportSpecimenData.class, "Import Specimen Data");
-        addAction(actionId, SpecimenController.ImportSpecimenData.class, "Import Specimen Data", directory, files, true, false, includeAll);
+        String actionId = createActionId(ImportSpecimenDataAction.class, "Import Specimen Data");
+        addAction(actionId, ImportSpecimenDataAction.class, "Import Specimen Data", directory, files, true, false, includeAll);
     }
-
 }
