+++ specimen/src/org/labkey/specimen/pipeline/SpecimenReloadJob.java
@@ -19,7 +19,6 @@ import org.labkey.api.pipeline.PipeRoot;
 import org.labkey.api.pipeline.PipelineJobService;
 import org.labkey.api.pipeline.TaskId;
 import org.labkey.api.pipeline.TaskPipeline;
-import org.labkey.api.specimen.pipeline.SpecimenBatch;
 import org.labkey.api.study.SpecimenTransform;
 import org.labkey.api.util.FileUtil;
 import org.labkey.api.view.ViewBackgroundInfo;
