+++ study/src/org/labkey/study/controllers/designer/DesignerController.java
@@ -49,6 +49,7 @@ import org.labkey.api.security.permissions.DeletePermission;
 import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.security.permissions.UpdatePermission;
 import org.labkey.api.specimen.importer.SimpleSpecimenImporter;
+import org.labkey.api.study.MapArrayExcelWriter;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.TimepointType;
@@ -65,7 +66,6 @@ import org.labkey.api.view.VBox;
 import org.labkey.api.view.template.ClientDependency;
 import org.labkey.study.controllers.BaseStudyController;
 import org.labkey.study.designer.JSONSerializer;
-import org.labkey.study.designer.MapArrayExcelWriter;
 import org.labkey.study.designer.StudyDefinitionServiceImpl;
 import org.labkey.study.designer.StudyDesignInfo;
 import org.labkey.study.designer.StudyDesignManager;
